fix: use cache index to find dataset path in scr_copy
