+++ src/scr_cache_index.c
@@ -289,70 +289,3 @@ int scr_cache_index_merge(scr_cache_index* cindex1, scr_cache_index* cindex2)
   kvtree_merge(cindex1, cindex2);
   return SCR_SUCCESS;
 }
-
-/* reads specified file and fills in cache index structure */
-int scr_cache_index_read(const spath* path_file, scr_cache_index* cindex)
-{
-  /* check that we have a cindex pointer and a hash within the cindex */
-  if (cindex == NULL) {
-    return SCR_FAILURE;
-  }
-
-  /* assume we'll fail */
-  int rc = SCR_FAILURE;
-
-  /* can't read file, return error (special case so as not to print error message below) */
-  if (scr_storedesc_cntl->rank == 0) {
-    /* get file name */
-    char* file = spath_strdup(path_file);
-
-    /* attempt to read the file */
-    if (scr_file_is_readable(file) == SCR_SUCCESS) {
-      /* ok, now try to read the file */
-      if (kvtree_read_file(file, cindex) == KVTREE_SUCCESS) {
-        /* successfully read the cache index file */
-        rc = SCR_SUCCESS;
-      } else {
-        scr_err("Reading cache index %s @ %s:%d",
-          file, __FILE__, __LINE__
-        );
-      }
-    }
-
-    /* free file name string */
-    scr_free(&file);
-  }
-
-  /* tell others whether rank 0 read the file */
-  MPI_Bcast(&rc, 1, MPI_INT, 0, scr_storedesc_cntl->comm);
-
-  /* bcast data to other ranks sharing the control directory */
-  if (rc == SCR_SUCCESS) {
-    kvtree_bcast(cindex, 0, scr_storedesc_cntl->comm);
-  }
-
-  return rc;
-}
-
-/* writes given cache index to specified file */
-int scr_cache_index_write(const spath* file, const scr_cache_index* cindex)
-{
-  /* check that we have a cindex pointer */
-  if (cindex == NULL) {
-    return SCR_FAILURE;
-  }
-
-  if (scr_storedesc_cntl->rank == 0) {
-    /* write out the hash */
-    if (kvtree_write_path(file, cindex) != KVTREE_SUCCESS) {
-      char path_err[SCR_MAX_FILENAME];
-      spath_strcpy(path_err, sizeof(path_err), file);
-      scr_err("Writing cache index %s @ %s:%d",
-        path_err, __FILE__, __LINE__
-      );
-      return SCR_FAILURE;
-    }
-  }
-
-  return SCR_SUCCESS;
-}
